+++ views/assignment/single.php
@@ -15,20 +15,24 @@
 
             <tbody>
                 <?php
-                    foreach ($data as $assignment) {
-                        echo '
-                        <tr>
-                        <td>'.$assignment["id"].'</td>
-                        <td><a href="'.BASEPATH."assignment/".$assignment['id'].'">'.$assignment["name"].'</a></td>
-                        <td>'.$assignment["deskripsi"].'</td>
-                        <td>'.$assignment["deadline"].'</td>
-                        <td>
-                            <a href="'.BASEPATH."assignment/".$assignment['id'].'/create" class="btn btn-sm btn-primary">add submission</a>
-                            <a href="'.BASEPATH."assignment/".$assignment['id'].'/edit" class="btn btn-sm btn-primary">edit</a>
-                            <a href="'.BASEPATH."assignment/".$assignment['id'].'/delete" class="btn btn-sm btn-danger">delete</a>
-                        </td>
-                        </tr>
-                        ';
+                    if (empty($data)) {
+                        echo '<div class="d-flex flex-column justify-content-center align-items-center display-1" style="min-height:100%;height:100vh;width:100%">Empty</div>'; 
+                    }else {
+                        foreach ($data as $assignment) {
+                            echo '
+                            <tr>
+                            <td>'.$assignment["id"].'</td>
+                            <td><a href="'.BASEPATH."assignment/".$assignment['id'].'">'.$assignment["name"].'</a></td>
+                            <td>'.$assignment["deskripsi"].'</td>
+                            <td>'.$assignment["deadline"].'</td>
+                            <td>
+                                <a href="'.BASEPATH."assignment/".$assignment['id'].'/create" class="btn btn-sm btn-primary">add submission</a>
+                                <a href="'.BASEPATH."assignment/".$assignment['id'].'/edit" class="btn btn-sm btn-primary">edit</a>
+                                <a href="'.BASEPATH."assignment/".$assignment['id'].'/delete" class="btn btn-sm btn-danger">delete</a>
+                            </td>
+                            </tr>
+                            ';
+                        }
                     }
                 ?>
             </tbody>
